Fixed bug caused by the previous memory leak fix.
The cause of the memory leak was not the "new" operator but rather the fact that the
are 2 firings of the velodyne and only in the second one we added the
points to the points cloud and subsequently deleted them. That left the
points from the first firing as a memory leak.
As a consequence of this fix, the point cloud is now denser.